split play.js
