+++ frontend/game/helper.js
@@ -2,9 +2,6 @@
 
 let params = new URLSearchParams(document.location.search.substring(1));
 
-const gameID = params.get("gameID");
-const playerID = params.get("playerID");
-
 define(function() {
     return {
         clearChildren(elem) {
@@ -13,22 +10,9 @@ define(function() {
             }
         },
         clickedCard: undefined,
-        gameID: gameID,
-        playerID: playerID,
+        gameID: params.get("gameID"),
+        playerID: params.get("playerID"),
         baseCard: document.getElementById("card-template")
-                .content.querySelector(".card"),
-        playCard(rownum, cardindex) {
-            var req = new XMLHttpRequest();
-            req.open("POST", `http://backend:3000/games/${gameID}/player/${playerID}`);
-            req.setRequestHeader("Content-type", "application/json");
-            req.responseType = "json";
-
-            const playData = {
-                index: cardindex,
-                row: rownum,
-            };
-
-            req.send(JSON.stringify(playData));
-        }
+                .content.querySelector(".card")
     }
-})
+});
